explain why implement a UNIX exec is hard in JOS
Copied from the comments I wrote within it:

Well, if we're going to implement some routine like 'exec', one big prerequisite is that we need to make sure we run under current process's context. Then it needs to find someway to replace itself with new binary image file, which means we have to be in the kernel mode to do that, otherwise how to maintain the correctness during the procedure of replacing? Then we need to of course update our page tables with newly added physical memory pages for new image, and also be sure to free the old image's physical memory occupation